+++ man-roxygen/elementStatusParam.R
@@ -1,0 +1,3 @@
+#' @param elementStatus Two-dimensional integer array, with rows corresponding to 
+#'   counts of matching quartets or partitions for each tree, and columns named 
+#'   according to the output of [QuartetStatus] or [SplitStatus].