+++ tests/testthat/test-QuartetDistance.R
@@ -4,7 +4,7 @@ data('sq_trees')
 quartets <- read.tree('../trees/all_quartets.new')
 
 test_that("Splits are compared", {
-  expect_equal(c(N=126, s=120, d=6, r1=0, r2=0, u=0), 
+  expect_equal(c(N=252L, Q=126, s=120, d=6, r1=0, r2=0, u=0), 
                SharedQuartetStatus(UnshiftTree(
                  ape::drop.tip(sq_trees$move_one_near, 10),
                  ape::drop.tip(sq_trees$ref_tree, 11)))[2, ])
@@ -24,7 +24,7 @@ test_that("QuartetState works", {
 })
 
 test_that("CompareQuartets works", {
-  expect_equal(c(N=8, s=1, d=2, r1=1, r2=1, u=3),
+  expect_equal(c(N=16L, Q=8L, s=1, d=2, r1=1, r2=1, u=3),
                CompareQuartets(c(2, 2, 4, 2, 0, 0, 0, 0),
                                c(2, 3, 3, 0, 2, 0, 0, 0)))
 })